+++ src/main/java/tn/esprit/pidev/repositories/RecuiterRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import tn.esprit.pidev.entities.Recuiter;
 
 @Repository
-public interface RecuiterRepository extends JpaRepository<Recuiter, Long> {
+public interface RecuiterRepository extends JpaRepository<Recuiter, Integer> {
 
 
 }
